Consolidate user gallery and single view
