+++ package.json
@@ -42,7 +42,7 @@
     "mapbox-gl": "^2.7.0",
     "nanoid": "^4.0.0",
     "nanoid-dictionary": "^4.3.0",
-    "next": "^12.1.7-canary.23",
+    "next": "^12.1.7-canary.45",
     "next-auth": "^4.5.0",
     "nprogress": "^0.2.0",
     "path-browserify": "^1.0.1",
@@ -57,7 +57,6 @@
     "react-hotkeys-hook": "^3.4.6",
     "react-map-gl": "^7.0.10",
     "react-paginate": "^8.1.3",
-    "react-placeholder": "^4.1.0",
     "react-popper": "^2.3.0",
     "react-responsive": "^9.0.0-beta.6",
     "react-swipeable": "^7.0.0",
@@ -152,8 +151,7 @@
     "files": [
       "src/**/*.js",
       "src/**/*.ts",
-      "src/**/*.tsx",
-      "src/pages/u/[uid]/[postId].tsx"
+      "src/**/*.tsx"
     ]
   }
 }
